Add keypair security level parameters to secret-key checksum
It may be useful to use the same field and matrix size for more than one
keypair security level.  Since the seeds contained in the key blobs are
manually sized, this would create a risk of using a keypair intended for
Category 2 security with Category 1 seeds; adding the blinding seed length
to the checksum plugs that hole.